u64 already holds up to 18 quintillion ms, which is 570 million years

+++ src/adapters/component_events.rs
@@ -6,7 +6,7 @@ pub enum ComponentEvent {
         username_component_name: String,
         stdout: Option<MemoryOutputPipe>,
         start: DateTime<chrono::Utc>,
-        duration: u128,
+        duration: u64,
         status: u16
     }
 }
